% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_bootstrap.R
\name{coef.fpgs_bootstrap}
\alias{coef.fpgs_bootstrap}
\title{Extract FPGS bootstrap estimate}
\usage{
\method{coef}{fpgs_bootstrap}(object, ...)
}
\arguments{
\item{object}{An object of class "fpgs_bootstrap".}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Extract FPGS bootstrap estimate
}
