% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_bootstrap.R
\name{confint.fpgs_bootstrap}
\alias{confint.fpgs_bootstrap}
\title{Confidence interval for FPGS bootstrap inference}
\usage{
\method{confint}{fpgs_bootstrap}(object, parm = NULL, level = object$conf.level, ...)
}
\arguments{
\item{object}{An object of class "fpgs_bootstrap".}

\item{parm}{Currently ignored.}

\item{level}{Confidence level. Defaults to the level stored in object.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Confidence interval for FPGS bootstrap inference
}
