% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_all.R
\name{fpgs_all}
\alias{fpgs_all}
\title{Run All FPGS Estimators}
\usage{
fpgs_all(fit)
}
\arguments{
\item{fit}{An object returned by fpgs()}
}
\value{
Data frame containing ATE estimates
}
\description{
Run All FPGS Estimators
}
