% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_binary_rf.R
\name{fpgs_binary_rf}
\alias{fpgs_binary_rf}
\title{Estimate Binary FPGS using Random Forest with Cross-Fitting}
\usage{
fpgs_binary_rf(
  data,
  outcome,
  treatment,
  covariates = NULL,
  folds = 5,
  num.trees = 300,
  mtry = NULL,
  min.node.size = 5
)
}
\arguments{
\item{data}{Data frame}

\item{outcome}{Binary outcome variable name}

\item{treatment}{Treatment variable name}

\item{covariates}{Covariate names}

\item{folds}{Number of folds}
}
\description{
Estimate Binary FPGS using Random Forest with Cross-Fitting
}
