% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_conti_par .R
\name{fpgs_continuous_parametric}
\alias{fpgs_continuous_parametric}
\title{Estimate Full Prognostic Scores}
\usage{
fpgs_continuous_parametric(data, outcome, treatment, covariates = NULL)
}
\arguments{
\item{data}{Data frame}

\item{outcome}{Outcome variable name}

\item{treatment}{Treatment variable name}

\item{covariates}{Covariate names}
}
\value{
An object of class "fpgs"
}
\description{
Estimates the prognostic scores
E\link{Y(0)|X} and E\link{Y(1)|X}
}
