% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_match.R
\name{fpgs_match}
\alias{fpgs_match}
\title{FPGS Matching Estimator}
\usage{
fpgs_match(fit, M = 1)
}
\arguments{
\item{fit}{An object returned by fpgs()}

\item{M}{Number of matches. Default is 1.}
}
\value{
An object of class "fpgs_match"
}
\description{
Estimates the ATE by matching on the two-dimensional FPGS:
mu0_hat and mu1_hat.
}
