% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_rf.R
\name{fpgs_rf}
\alias{fpgs_rf}
\title{Estimate FPGS using random forest with cross-fitting}
\usage{
fpgs_rf(
  data,
  outcome,
  treatment,
  covariates = NULL,
  folds = 5,
  num.trees = 500,
  mtry = NULL,
  min.node.size = 5
)
}
\arguments{
\item{data}{Data frame}

\item{outcome}{Outcome variable name}

\item{treatment}{Treatment variable name}

\item{covariates}{Covariate names}

\item{folds}{Number of folds}
}
\description{
Estimate FPGS using random forest with cross-fitting
}
