% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_ri_est.R
\name{fpgs_ri}
\alias{fpgs_ri}
\title{FPGS Regression Imputation Estimator}
\usage{
fpgs_ri(fit)
}
\arguments{
\item{fit}{An object returned by fpgs()}
}
\value{
An object of class "fpgs_ri"
}
\description{
FPGS Regression Imputation Estimator
}
