% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_ipw.R
\name{fpgs_weight}
\alias{fpgs_weight}
\title{FPGS Weighting Estimator}
\usage{
fpgs_weight(fit, type = c("ipw", "normalized", "aipw"))
}
\arguments{
\item{fit}{An object returned by fpgs()
#' @param type Weighting estimator: "ipw", "normalized", or "aipw"}
}
\description{
Estimates ATE using propensity scores estimated from FPGS.
}
