% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_bootstrap.R
\name{print.fpgs_bootstrap}
\alias{print.fpgs_bootstrap}
\title{Print method for FPGS bootstrap inference}
\usage{
\method{print}{fpgs_bootstrap}(x, ...)
}
\arguments{
\item{x}{An object of class "fpgs_bootstrap".}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Print method for FPGS bootstrap inference
}
