% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_bootstrap.R
\name{summary.fpgs_bootstrap}
\alias{summary.fpgs_bootstrap}
\title{Summary method for FPGS bootstrap inference}
\usage{
\method{summary}{fpgs_bootstrap}(object, ...)
}
\arguments{
\item{object}{An object of class "fpgs_bootstrap".}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Summary method for FPGS bootstrap inference
}
