% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_bootstrap.R
\name{vcov.fpgs_bootstrap}
\alias{vcov.fpgs_bootstrap}
\title{Variance-covariance matrix for FPGS bootstrap estimate}
\usage{
\method{vcov}{fpgs_bootstrap}(object, ...)
}
\arguments{
\item{object}{An object of class "fpgs_bootstrap".}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Variance-covariance matrix for FPGS bootstrap estimate
}
